+++ lib/keycloak-admin/client/role_client.rb
@@ -12,11 +12,18 @@ module KeycloakAdmin
       end
       JSON.parse(response).map { |role_as_hash| RoleRepresentation.from_hash(role_as_hash) }
     end
+    
+    def get(name)
+      response = execute_http do
+        RestClient::Resource.new(role_name_url(name), @configuration.rest_client_options).get(headers)
+      end
+      RoleRepresentation.from_hash JSON.parse(response)
+    end
 
     def save(role_representation)
       execute_http do
         RestClient::Resource.new(roles_url, @configuration.rest_client_options).post(
-          role_representation.to_json, headers
+          create_payload(role_representation), headers
         )
       end
     end
@@ -28,5 +35,10 @@ module KeycloakAdmin
         "#{@realm_client.realm_admin_url}/roles"
       end
     end
+    
+    def role_name_url(name)
+      "#{@realm_client.realm_admin_url}/roles/#{name}"
+    end
+    
   end
 end