+++ spec/client/client_client_spec.rb
@@ -22,6 +22,49 @@ RSpec.describe KeycloakAdmin::ClientClient do
     end
   end
 
+  describe "#get" do
+    let(:realm_name) { "valid-realm" }
+    let(:id) { "test_client_id" }
+    let(:client_name) { "test_client_name" }
+
+    before(:each) do
+      @client_client = KeycloakAdmin.realm(realm_name).clients
+
+      stub_token_client
+      allow_any_instance_of(RestClient::Resource).to receive(:get).and_return '{"id":"test_client_id","name":"test_client_name"}'
+    end
+
+    it "finds a client" do
+      client = @client_client.get(id)
+      expect(client.name).to eq client_name
+      expect(client.id).to eq id
+    end
+  end
+
+  describe "#find_by_client_id" do
+    let(:realm_name) { "valid-realm" }
+    let(:client_id) { "my_client_id" }
+    let(:client_name) { "test_client_name" }
+
+    before(:each) do
+      @client_client = KeycloakAdmin.realm(realm_name).clients
+
+      stub_token_client
+      allow_any_instance_of(RestClient::Resource).to receive(:get).and_return '[{"id":"test_client_id","clientId": "my_client_id","name":"test_client_name"},{"id":"test_client_id_2","clientId":"client_id_2","name":"test_client_name_2"}]'
+    end
+
+    it "finds a client it has" do
+      client = @client_client.find_by_client_id(client_id)
+      expect(client.name).to eq client_name
+      expect(client.client_id).to eq client_id
+    end
+
+    it "returns nil if it doesn't have the client" do
+      client = @client_client.find_by_client_id("client_id_3")
+      expect(client).to be_nil
+    end
+  end
+
   describe "#list" do
     let(:realm_name) { "valid-realm" }
 
@@ -48,6 +91,43 @@ RSpec.describe KeycloakAdmin::ClientClient do
       clients = @client_client.list
       expect(clients.length).to eq 1
       expect(clients[0].name).to eq "test_client_name"
+    end
+  end
+
+  describe "#update" do
+    let(:realm_name) { "valid-realm" }
+    let(:client) { KeycloakAdmin::ClientRepresentation.from_hash({ "id" => "test_client_id", "clientId" => "my-client", "name" => "old_name" }) }
+
+    before(:each) do
+      @client_client = KeycloakAdmin.realm(realm_name).clients
+
+      stub_token_client
+      allow_any_instance_of(RestClient::Resource).to receive(:put).and_return ''
+      allow_any_instance_of(RestClient::Resource).to receive(:get).and_return '{"id":"test_client_id", "clientId": "my-client","name":"new_name"}'
+    end
+
+    it "updates a client" do
+      updated_client = @client_client.update(client)
+
+      expect(updated_client.name).to eq "new_name"
+    end
+  end
+
+  describe "#delete" do
+    let(:realm_name) { "valid-realm" }
+
+    before(:each) do
+      @client_client = KeycloakAdmin.realm(realm_name).clients
+      stub_token_client
+      allow_any_instance_of(RestClient::Resource).to receive(:delete).and_return ''
+    end
+
+    it "passes rest client options" do
+      rest_client_options = {verify_ssl: OpenSSL::SSL::VERIFY_NONE}
+      allow_any_instance_of(KeycloakAdmin::Configuration).to receive(:rest_client_options).and_return rest_client_options
+      expect(RestClient::Resource).to receive(:new).with(
+        "http://auth.service.io/auth/admin/realms/valid-realm/clients/95b45037-3980-404c-ba12-784fa1baf2c2", rest_client_options).and_call_original
+      @client_client.delete("95b45037-3980-404c-ba12-784fa1baf2c2")
     end
   end
 end