+++ spec/spec_helper.rb
@@ -10,7 +10,7 @@ def configure
     config.client_secret       = "aaaaaaaa"
     config.client_realm_name   = "master2"
     config.use_service_account = true
-  end 
+  end
 end
 
 RSpec.configure do |config|
@@ -23,7 +23,15 @@ end
 
 def stub_token_client
   allow_any_instance_of(KeycloakAdmin::TokenClient).to receive(:get).and_return KeycloakAdmin::TokenRepresentation.new(
-    'test_access_token', 'token_type', 'expires_in', 'refresh_token',
-    'refresh_expires_in', 'id_token', 'not_before_policy', 'session_state'
+    "test_access_token", "token_type", "expires_in", "refresh_token",
+    "refresh_expires_in", "id_token", "not_before_policy", "session_state"
   )
+end
+
+def stub_net_http_res(res_class, code, message)
+  net_http_res = double(message:  message, code: code)
+  allow(net_http_res).to receive(:is_a?) do |target_class|
+    target_class == res_class
+  end
+  allow(@response).to receive(:net_http_res).and_return(net_http_res)
 end